refactor(fn): renamed files to dash-case part 4
