+++ libs/fn/src/lib/utils/index.ts
@@ -1,8 +1,8 @@
 export * from './utils.module';
-export * from './decorators/coerceBoolean';
+export * from './decorators/coerce-boolean';
 export * from './template-ref.directive';
 export * from './template-ref-provider.token';
 export * from './utils.tokens';
 export * from './utils.types';
 export * from './helpers/get-native-element';
-export * from './helpers/pxToNum';
+export * from './helpers/px-to-num';
