+++ libs/fn/src/lib/utils/utils.module.ts
@@ -1,9 +1,9 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule, DOCUMENT } from '@angular/common';
 import { GetComputedStyle, GetElementCapacity, GetElementWidth } from './utils.tokens';
-import { getComputedStyleFactory } from './factories/getComputedStyleFactory';
-import { getElementWidthFactory } from './factories/getElementWidthFactory';
-import { getElementCapacityFactory } from './factories/getElementCapacityFactory';
+import { getComputedStyleFactory } from './factories/get-computed-style.factory';
+import { getElementWidthFactory } from './factories/get-element-width.factory';
+import { getElementCapacityFactory } from './factories/get-element-capacity.factory';
 
 @NgModule({
     imports: [CommonModule],
